refactor: extract feedback-record access and error mapping
operations.ts had grown to hold three concerns: authorization, Hub error
mapping, and the operations themselves. Split along the lines the rest of
the v3 layer already uses (see unify-feedback/taxonomy/lib/access.ts):

- lib/access.ts — the workspace/dataset tenant choke point, plus a new
  requireOwnedFeedbackRecord() that factors out the retrieve-then-assert
  guard get_feedback_record had inline. The Hub's get/delete/similar
  endpoints all take a bare record id and derive the tenant from the
  record, so each of them needs exactly this check.
- lib/errors.ts — Hub error to v3 problem mapping, with its own tests for
  the relay bounds that cap what a remote service can put in our body.

No behaviour change.
